Fix memory limits
If a user requested a node with default options (standard), they could
receive a bigmem node. The scripts would configure Spark based on the
amount of memory in the node, which could be significantly higher than
what Slurm actually allows. Job were getting OOM-killed as a result.
Look at the cgroup settings instead.